+++ lbrynet/core/Wallet.py
@@ -17,6 +17,7 @@ from lbryum.network import Network
 from lbryum.simple_config import SimpleConfig
 from lbryum.constants import COIN
 from lbryum.commands import Commands
+from lbryum.errors import InvalidPassword
 
 from lbryschema.uri import parse_lbry_uri
 from lbryschema.claim import ClaimDict
@@ -1135,6 +1136,9 @@ class LBRYumWallet(Wallet):
         self.config = make_config(self._config)
         self.network = None
         self.wallet = None
+        self._cmd_runner = None
+        self.wallet_pw_d = None
+        self.wallet_unlocked_d = defer.Deferred()
         self.is_first_run = False
         self.printed_retrieving_headers = False
         self._start_check = None
@@ -1169,6 +1173,37 @@ class LBRYumWallet(Wallet):
             else:
                 network_start_d.errback(ValueError("Failed to connect to network."))
 
+        def get_cmd_runner():
+            self._cmd_runner = Commands(self.config, self.wallet, self.network)
+
+        def unlock(password):
+            if self._cmd_runner and self._cmd_runner.locked:
+                try:
+                    self._cmd_runner.unlock_wallet(password)
+                except InvalidPassword:
+                    log.warning("Incorrect password")
+                    check_locked()
+                    raise InvalidPassword
+            if self._cmd_runner and self._cmd_runner.locked:
+                raise Exception("Failed to unlock wallet")
+            elif not self._cmd_runner:
+                raise Exception("Command runner hasn't been initialized yet")
+            self.wallet_unlocked_d.callback(True)
+            log.info("Unlocked the wallet!")
+
+        def check_locked():
+            if self._cmd_runner and self._cmd_runner.locked:
+                log.info("Waiting for wallet password")
+                d = defer.Deferred()
+                d.addCallback(unlock)
+                self.wallet_pw_d = d
+                return self.wallet_unlocked_d
+            elif not self._cmd_runner:
+                raise Exception("Command runner hasn't been initialized yet")
+            if not self.wallet.use_encryption:
+                log.info("Wallet is not encrypted")
+            self.wallet_unlocked_d.callback(True)
+
         self._start_check = task.LoopingCall(check_started)
 
         d = setup_network()
@@ -1180,6 +1215,9 @@ class LBRYumWallet(Wallet):
         d.addCallback(lambda _: log.info("Subscribing to addresses"))
         d.addCallback(lambda _: self.wallet.wait_until_synchronized(lambda _: None))
         d.addCallback(lambda _: log.info("Synchronized wallet"))
+        d.addCallback(lambda _: get_cmd_runner())
+        d.addCallbacks(lambda _: log.info("Set up lbryum command runner"))
+        d.addCallbacks(lambda _: check_locked())
         return d
 
     def _stop(self):
@@ -1261,16 +1299,12 @@ class LBRYumWallet(Wallet):
         d.addCallback(lambda _: blockchain_caught_d)
         return d
 
-    def _get_cmd_runner(self):
-        return Commands(self.config, self.wallet, self.network)
-
     # run commands as a defer.succeed,
     # lbryum commands should be run this way , unless if the command
     # only makes a lbrum server query, use _run_cmd_as_defer_to_thread()
     def _run_cmd_as_defer_succeed(self, command_name, *args, **kwargs):
-        cmd_runner = self._get_cmd_runner()
         cmd = Commands.known_commands[command_name]
-        func = getattr(cmd_runner, cmd.name)
+        func = getattr(self._cmd_runner, cmd.name)
         return defer.succeed(func(*args, **kwargs))
 
     # run commands as a deferToThread,  lbryum commands that only make
@@ -1278,9 +1312,8 @@ class LBRYumWallet(Wallet):
     # TODO: keep track of running threads and cancel them on `stop`
     #       otherwise the application will hang, waiting for threads to complete
     def _run_cmd_as_defer_to_thread(self, command_name, *args, **kwargs):
-        cmd_runner = self._get_cmd_runner()
         cmd = Commands.known_commands[command_name]
-        func = getattr(cmd_runner, cmd.name)
+        func = getattr(self._cmd_runner, cmd.name)
         return threads.deferToThread(func, *args, **kwargs)
 
     def _update_balance(self):