+++ CHANGELOG.md
@@ -12,12 +12,14 @@ at anytime.
   * Add link to instructions on how to change the default peer port
   * Add `peer_port` to settings configurable using `settings_set`
   * Added an option to disable max key fee check.
+  * Add `wallet_unlock`, a command available during startup to unlock an encrypted wallet
 
 ### Changed
   * claim_show API command no longer takes name as argument
   * Linux default downloads folder changed from `~/Downloads` to `XDG_DOWNLOAD_DIR`
   * Linux folders moved from the home directory to `~/.local/share/lbry`
   * Windows folders moved from `%AppData%/Roaming` to `%AppData%/Local/lbry`
+  * Block wallet startup on being unlocked
 
 ### Fixed
   * Fix for https://github.com/lbryio/lbry/issues/750